requirements.txt
pyhomebroker==0.55
xlwings==0.30.12
django-environ==0.4.5

broker=153
dni=26386662
user=piolifl
password=Piolifl1511
account_id=2474
name=BCCH

broker=265
dni=26386662
user=piolifl
password=Piolifl1511
account_id=10214
name=COCOS.CAPITAL

broker=265
dni=26968339
user=bordame
password=Bordame03
account_id=49765
name=COCOS.CAPITAL



